+++ nengo/utils/nco.py
@@ -25,6 +25,12 @@ array data an alignment of 16 bytes.
 
 The Numpy NPY format is documented here:
 https://github.com/numpy/numpy/blob/master/doc/neps/npy-format.rst
+
+As of legacy version 1 of the cache, multiple NCO files will be concatenated
+into one file. The start and end of each subfile will be stored in a cache
+index, but can also be recovered from reading the headers of the NCO files in
+order as each one gives the start of the next header (corresponding to the
+end of the array data).
 """
 
 from __future__ import absolute_import
@@ -42,7 +48,7 @@ from ..exceptions import CacheIOError
 class Subfile(object):
     """A file-like object for limiting reads to a subrange of a file.
 
-    This clss only supports reading and seeking. Writing is not supported.
+    This class only supports reading and seeking. Writing is not supported.
 
     Parameters
     ----------
@@ -85,6 +91,9 @@ class Subfile(object):
         self.max_size = self.end - offset
         self.fileobj.seek(offset)
 
+    def tell(self):
+        return self.fileobj.tell() - self.start
+
 
 MAGIC_STRING = ensure_bytes('NCO')
 SUPPORTED_PROTOCOLS = [0]
@@ -105,7 +114,8 @@ def write(fileobj, metadata, array):
     array : ndarray
         Numpy array with the actual data to store.
     """
-    pickle_start = byte_align(HEADER_SIZE, ALIGNMENT)
+    start = fileobj.tell()
+    pickle_start = byte_align(start + HEADER_SIZE, ALIGNMENT)
     fileobj.seek(pickle_start)
     pickle.dump(metadata, fileobj, pickle.HIGHEST_PROTOCOL)
     pickle_end = fileobj.tell()
@@ -118,8 +128,9 @@ def write(fileobj, metadata, array):
     header = struct.pack(
         HEADER_FORMAT, MAGIC_STRING, 0, pickle_start, pickle_end,
         array_start, array_end)
-    fileobj.seek(0)
+    fileobj.seek(start)
     fileobj.write(header)
+    fileobj.seek(array_end)
 
 
 def read(fileobj):